pyproject.toml
[project]
name = "fastapi-todo"
version = "0.1.0"
description = "Todo with FastAPI"
authors = [
    {name = "Anton Os"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.128.8,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "redis (>=7.1.1,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.0,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
